Budget details: extracted description input to own component
